support unseekable streams: skip encoding detection
DetectEncodingOrNull requires a seekable stream. Encoding detection
requires reading the first four bytes. The stream is advanced, then
rewound. Fine for a FileStream, but not for a DeflateStream, as when the
xml is being read with System.IO.Compression.ZipArchive.

SO. If the caller has provided an encoding: use it, skip trying to suss
out the actual stream encoding, and absolutely avoid writing to the
position property of the stream.